split ContentView inner views into separate files/structs
