+++ chronicle/Views/EnrollmentSuccessMessage.swift
@@ -1,0 +1,24 @@
+//
+//  EnrollmentSuccessMessage.swift
+//  EnrollmentSuccessMessage
+//
+//  Created by Giulia Campana on 8/18/21.
+//
+
+import SwiftUI
+
+struct EnrollmentSuccessMessage: View {
+    var body: some View {
+        VStack(alignment: .leading) {
+            Text("Your device has been enrolled! Thank you for participating in our study!\n\nYou have successfully installed Chronicle and may now close the application. Chronicle will continue to function as a background process on your device while the application is closed. To stop sending information to the researchers - for instance, at the end of the study - simply uninstall this app. If you have questions or concerns, please contact a member of your research team.")
+            // later, add "Done" button which navigates to home page
+            Spacer()
+        }.padding()
+    }
+}
+
+struct EnrollmentSuccessMessage_Previews: PreviewProvider {
+    static var previews: some View {
+        EnrollmentSuccessMessage()
+    }
+}
